requirements.txt
opencv-python==4.12.0.88
requests==2.32.4
numpy==2.2.*
pytz==2025.2
pillow==11.3.0